Move rack app logic to Wren::App
